+++ src/hyperion-qt/hyperion-qt.cpp
@@ -15,6 +15,7 @@
 
 // ssdp discover
 #include <ssdp/SSDPDiscover.h>
+#include <utils/NetUtils.h>
 
 using namespace commandline;
 
@@ -126,8 +127,17 @@ int main(int argc, char ** argv)
 				}
 			}
 
+			// Resolve hostname and port (or use default port)
+			QString host;
+			quint16 port { FLATBUFFER_DEFAULT_PORT };
+
+			if ( !NetUtils::resolveHostPort(address, host, port) )
+			{
+				throw std::runtime_error(QString("Wrong address: unable to parse address (%1)").arg(address).toStdString());
+			}
+
 			// Create the Flabuf-connection
-			FlatBufferConnection flatbuf("Qt Standalone", address, argPriority.getInt(parser), parser.isSet(argSkipReply));
+			FlatBufferConnection flatbuf("Qt Standalone", host, argPriority.getInt(parser), parser.isSet(argSkipReply), port);
 
 			// Connect the screen capturing to flatbuf connection processing
 			QObject::connect(&qtWrapper, SIGNAL(sig_screenshot(const Image<ColorRgb> &)), &flatbuf, SLOT(setImage(Image<ColorRgb>)));