+++ src/hyperion-remote/JsonConnection.cpp
@@ -18,45 +18,17 @@
 // util includes
 #include <utils/JsonUtils.h>
 
-// Constants
-namespace {
-const int JSON_DEFAULT_PORT = 19444;
-} //End of constants
-
-JsonConnection::JsonConnection(const QString & address, bool printJson)
+JsonConnection::JsonConnection(const QString & host, bool printJson , quint16 port)
 	: _printJson(printJson)
 	, _log(Logger::getInstance("REMOTE"))
 {
-	// Resolve hostname and port (or use default JSON-port)
-	QString testUrl;
-	if (address.at(0) != '[' && address.count(':') > 1)
-	{
-		testUrl = QString("http://[%1]").arg(address);
-	}
-	else
-	{
-		testUrl = QString("http://%1").arg(address);
-	}
-
-	QUrl url(testUrl);
-	if (!url.isValid())
-	{
-		throw std::runtime_error(QString("Wrong address: unable to parse address (%1)").arg(address).toStdString());
-	}
-
-	int port = url.port();
-	if (port == -1)
-	{
-		port = JSON_DEFAULT_PORT;
-	}
-
-	_socket.connectToHost(url.host(), static_cast<quint16>(port));
+	_socket.connectToHost(host, port);
 	if (!_socket.waitForConnected())
 	{
-		throw std::runtime_error(QString("Unable to connect to host (%1), port (%2)").arg(address).arg(port).toStdString());
+		throw std::runtime_error(QString("Unable to connect to host (%1), port (%2)").arg(host).arg(port).toStdString());
 	}
 
-	qDebug() << "Connected to:" << url.host() << "port:" << port;
+	qDebug() << "Connected to:" <<host << "port:" << port;
 }
 
 JsonConnection::~JsonConnection()