+++ src/hyperion-v4l2/hyperion-v4l2.cpp
@@ -23,6 +23,7 @@
 
 // ssdp discover
 #include <ssdp/SSDPDiscover.h>
+#include <utils/NetUtils.h>
 
 #include <utils/DefaultSignalHandler.h>
 
@@ -241,8 +242,17 @@ int main(int argc, char** argv)
 				}
 			}
 
-			// Create the Flatbuf-connection
-			FlatBufferConnection flatbuf("V4L2 Standalone", address, argPriority.getInt(parser), parser.isSet(argSkipReply));
+			// Resolve hostname and port (or use default port)
+			QString host;
+			quint16 port{ FLATBUFFER_DEFAULT_PORT };
+
+			if (!NetUtils::resolveHostPort(address, host, port))
+			{
+				throw std::runtime_error(QString("Wrong address: unable to parse address (%1)").arg(address).toStdString());
+			}
+
+			// Create the Flabuf-connection
+			FlatBufferConnection flatbuf("V4L2 Standalone", host, argPriority.getInt(parser), parser.isSet(argSkipReply), port);
 
 			// Connect the screen capturing to flatbuf connection processing
 			QObject::connect(&grabber, SIGNAL(newFrame(const Image<ColorRgb> &)), &flatbuf, SLOT(setImage(Image<ColorRgb>)));